+++ src/main/java/com/rackspace/ceres/app/services/MetadataService.java
@@ -51,6 +51,7 @@ import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
 import org.reactivestreams.Publisher;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
 import org.springframework.data.cassandra.core.ReactiveCassandraTemplate;
 import org.springframework.data.cassandra.core.cql.ReactiveCqlTemplate;
 import org.springframework.stereotype.Service;
@@ -60,12 +61,11 @@ import reactor.core.publisher.Mono;
 
 @Service
 @Slf4j
+@Profile({"ingest", "query", "admin"})
 public class MetadataService {
 
-  private static final String PREFIX_SERIES_SET_HASHES = "seriesSetHashes";
   private static final String LABEL_METRIC_GROUP = "metricGroup";
   private static final String LABEL_RESOURCE = "resource";
-  private static final String DELIM = "|";
   private static final String TAG_VALUE_REGEX = ".*\\{.*\\}$";
   private final ReactiveCqlTemplate cqlTemplate;
   private final ReactiveCassandraTemplate cassandraTemplate;
