+++ src/main/java/com/rackspace/ceres/app/services/MetricDeletionService.java
@@ -31,6 +31,7 @@ import java.util.List;
 import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
@@ -38,6 +39,7 @@ import reactor.core.publisher.Sinks.Empty;
 
 @Service
 @Slf4j
+@Profile({"ingest", "query", "admin"})
 public class MetricDeletionService {
 
   private final DataTablesStatements dataTablesStatements;
