+++ src/test/java/com/rackspace/ceres/app/services/QueryServiceTest.java
@@ -16,32 +16,14 @@
 
 package com.rackspace.ceres.app.services;
 
-import static org.apache.commons.lang3.RandomStringUtils.randomAlphanumeric;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.when;
-
 import com.rackspace.ceres.app.CassandraContainerSetup;
 import com.rackspace.ceres.app.config.DownsampleProperties;
 import com.rackspace.ceres.app.config.DownsampleProperties.Granularity;
 import com.rackspace.ceres.app.downsample.Aggregator;
 import com.rackspace.ceres.app.entities.MetricName;
 import com.rackspace.ceres.app.entities.SeriesSet;
-import com.rackspace.ceres.app.model.FilterType;
-import com.rackspace.ceres.app.model.Metric;
-import com.rackspace.ceres.app.model.MetricNameAndTags;
-import com.rackspace.ceres.app.model.PendingDownsampleSet;
-import com.rackspace.ceres.app.model.TsdbFilter;
-import com.rackspace.ceres.app.model.TsdbQuery;
-import com.rackspace.ceres.app.model.TsdbQueryRequest;
+import com.rackspace.ceres.app.model.*;
 import com.rackspace.ceres.app.utils.DateTimeUtils;
-import java.time.Duration;
-import java.time.Instant;
-import java.time.temporal.ChronoUnit;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.jupiter.api.AfterEach;
@@ -62,6 +44,18 @@ import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import reactor.test.StepVerifier;
 import reactor.util.function.Tuples;
+
+import java.time.Duration;
+import java.time.Instant;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
+
+import static org.apache.commons.lang3.RandomStringUtils.randomAlphanumeric;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.when;
 
 @SpringBootTest
 @ActiveProfiles(profiles = {"test", "downsample","query"})
@@ -96,6 +90,9 @@ class QueryServiceTest {
 
   @Autowired
   QueryService queryService;
+
+  @Autowired
+  DownsampleQueryService downsampleQueryService;
 
   @Autowired
   ReactiveCassandraTemplate cassandraTemplate;
@@ -235,59 +232,6 @@ class QueryServiceTest {
   }
 
   @Test
-  void testQueryRawWithSeriesSet() {
-    final String tenantId = RandomStringUtils.randomAlphanumeric(10);
-    final String metricGroup = RandomStringUtils.randomAlphabetic(5);
-    final String metricName = RandomStringUtils.randomAlphabetic(5);
-    final String resource = RandomStringUtils.randomAlphabetic(5);
-    final String monitoring_system = RandomStringUtils.randomAlphanumeric(5);
-    final Map<String, String> tags = Map.of(
-        "os", "linux",
-        "host", "h-1",
-        "deployment", "prod",
-        "metricGroup", metricGroup,
-        "resource", resource,
-        "monitoring_system", monitoring_system
-    );
-    final String seriesSetHash = seriesSetService
-        .hash(metricName, tags);
-
-    when(ingestTrackingService.track(any(), anyString(), any()))
-        .thenReturn(Mono.empty());
-
-    when(metadataService.storeMetadata(any(), any(), any()))
-        .thenReturn(Mono.empty());
-
-    when(metadataService.updateMetricGroupAddMetricName(anyString(), anyString(), any(), any()))
-        .thenReturn(Mono.empty());
-    when(metadataService.updateDeviceAddMetricName(anyString(), anyString(), any(), any()))
-        .thenReturn(Mono.empty());
-    when(metadataService.locateSeriesSetHashes(anyString(), anyString(), any()))
-        .thenReturn(Flux.just(seriesSetHash));
-
-    MetricNameAndTags metricNameAndTags = new MetricNameAndTags().setTags(tags).setMetricName(metricName);
-    when(metadataService.resolveSeriesSetHash(anyString(), anyString()))
-        .thenReturn(Mono.just(metricNameAndTags));
-
-    Instant instant = Instant.now();
-    dataWriteService.ingest(
-        tenantId,
-        new Metric()
-            .setTimestamp(instant)
-            .setValue(Math.random())
-            .setMetric(metricName)
-            .setTags(tags)
-    ).block();
-
-    StepVerifier.create(queryService
-            .queryRawWithSeriesSet(tenantId, seriesSetHash, Instant.now().minusSeconds(60), Instant.now()).collectList())
-        .assertNext(result -> {
-          assertThat(result).isNotEmpty();
-          assertThat(result.get(0).getTimestamp()).isEqualTo(instant.truncatedTo(ChronoUnit.MILLIS));
-        }).verifyComplete();
-  }
-
-  @Test
   void testQueryDownsampledWithMetricName() {
     final String tenant = randomAlphanumeric(10);
     final String metricName = RandomStringUtils.randomAlphabetic(5);
@@ -330,7 +274,7 @@ class QueryServiceTest {
         this.downsamplingService.downsampleData(
             pendingSet,
             granularity.getWidth(),
-            queryService.fetchData(pendingSet, group, granularity.getWidth(), false)
+            downsampleQueryService.fetchData(pendingSet, group, granularity.getWidth(), false)
         ).subscribe()
     );
 
@@ -387,7 +331,7 @@ class QueryServiceTest {
         this.downsamplingService.downsampleData(
             pendingSet,
             granularity.getWidth(),
-            queryService.fetchData(pendingSet, group, granularity.getWidth(), false)
+            downsampleQueryService.fetchData(pendingSet, group, granularity.getWidth(), false)
         ).subscribe()
     );
 
@@ -470,7 +414,7 @@ class QueryServiceTest {
             this.downsamplingService.downsampleData(
                 pendingSet.setTimeSlot(timeslot),
                 granularity.getWidth(),
-                queryService.fetchData(
+                downsampleQueryService.fetchData(
                     pendingSet.setTimeSlot(timeslot),
                     group, granularity.getWidth(), false)
             ).subscribe()
